requirements.txt
# === CORE FRAMEWORK ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiogram==3.8.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# === REDIS & ASYNC ===
redis==5.0.1

# === TASK QUEUE ===
celery[redis]==5.4.0

# === AI LIBRARIES (Fully supported in Python 3.12) ===
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
diffusers==0.25.0
transformers==4.36.2
accelerate==0.25.0
huggingface-hub==0.19.4

# pip install safetensors --only-binary=all


# === IMAGE PROCESSING ===
Pillow==10.1.0
opencv-python==4.9.0.80

# === PAYMENT SYSTEMS ===
yookassa==3.0.0
requests==2.31.0
aiohttp

# === SECURITY & UTILS ===
cryptography
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2

# === STORAGE & FILE HANDLING ===
minio==7.2.0
boto3==1.34.0

# === MONITORING & LOGGING ===
prometheus-client==0.19.0
structlog==23.2.0

# === DATABASE MIGRATION ===
psycopg2-binary==2.9.9

# === DEVELOPMENT TOOLS ===
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# === ENVIRONMENT & CONFIG ===
python-dotenv==1.0.0
click==8.1.7

# === ADDITIONAL DEPENDENCIES ===
typing-extensions==4.10.0
pyjwt==2.8.0
email-validator==2.1.0
python-magic==0.4.27
